requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
cycler==0.10.0
cymem==2.0.3
DateTime==4.3
Flask==1.1.2
idna==2.10
inflection==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
murmurhash==1.0.2
numpy==1.19.1
openapi-spec-validator==0.2.9
pandas==1.1.0
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
spacy==2.3.2
srsly==1.0.2
swagger-ui-bundle==0.0.8
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.48.0
urllib3==1.25.10
wasabi==0.7.1
Werkzeug==1.0.1
zope.interface==5.1.0
